+++ src/full/Agda/Termination/TermCheck.hs
@@ -686,6 +686,7 @@ instance ExtractCalls Sort where
       Inf f n    -> return empty
       SizeUniv   -> return empty
       LockUniv   -> return empty
+      IntervalUniv -> return empty
       Type t     -> terUnguarded $ extract t  -- no guarded levels
       Prop t     -> terUnguarded $ extract t
       SSet t     -> terUnguarded $ extract t