+++ src/full/Agda/TypeChecking/MetaVars/Mention.hs
@@ -55,6 +55,7 @@ instance MentionsMeta Sort where
     SSet l     -> mentionsMetas xs l
     SizeUniv   -> False
     LockUniv   -> False
+    IntervalUniv -> False
     PiSort a s1 s2 -> mentionsMetas xs (a, s1, s2)
     FunSort s1 s2 -> mentionsMetas xs (s1, s2)
     UnivSort s -> mentionsMetas xs s