+++ src/full/Agda/TypeChecking/Primitive/Cubical.hs
@@ -56,7 +56,7 @@ requireCubical wanted s = do
     CErased -> "--cubical or --erased-cubical"
 
 primIntervalType :: (HasBuiltins m, MonadError TCErr m, MonadTCEnv m, ReadTCState m) => m Type
-primIntervalType = El (SSet $ ClosedLevel 0) <$> primInterval
+primIntervalType = El IntervalUniv <$> primInterval
 
 primINeg' :: TCM PrimitiveImpl
 primINeg' = do
@@ -1082,7 +1082,14 @@ primTransHComp cmd ts nelims = do
                                                          <@> phi
                                                          <@> a0
           LockUniv -> return $ Just $ \ _ _ a0 -> a0
-          _       -> return Nothing
+          IntervalUniv -> do
+            x' <- reduceB $ unDom x
+            mInterval <- getBuiltinName' builtinInterval
+            case unEl $ ignoreBlocking x' of
+              Def q [] | Just q == mInterval -> return $ Just $ \ _ _ a0 -> a0
+              _ -> return Nothing
+          _ -> return Nothing
+
       caseMaybe labA (return Nothing) $ \ trA -> Just <$> do
       [phi, u0] <- mapM (open . unArg) [phi, u0]
       u <- traverse open (unArg <$> u)