+++ src/full/Agda/TypeChecking/Serialise/Instances/Internal.hs
@@ -153,6 +153,7 @@ instance EmbPrj I.Sort where
   icod_ (DefS a b)   = icodeN 7 DefS a b
   icod_ (SSet  a  ) = icodeN 8 SSet a
   icod_ LockUniv    = icodeN 9 LockUniv
+  icod_ IntervalUniv = icodeN 10 IntervalUniv
   icod_ (DummyS s)   = do
     liftIO $ putStrLn $ "Dummy sort in serialization: " ++ s
     __IMPOSSIBLE__
@@ -168,6 +169,7 @@ instance EmbPrj I.Sort where
     valu [7, a, b] = valuN DefS a b
     valu [8, a]    = valuN SSet a
     valu [9]       = valuN LockUniv
+    valu [10]      = valuN IntervalUniv
     valu _         = malformed
 
 instance EmbPrj DisplayForm where
@@ -262,7 +264,8 @@ instance EmbPrj NLPSort where
   icod_ (PProp a)   = icodeN 1 PProp a
   icod_ (PInf f a)  = icodeN 2 PInf f a
   icod_ PSizeUniv   = icodeN 3 PSizeUniv
-  icod_ PLockUniv   = icodeN 4 PSizeUniv
+  icod_ PLockUniv   = icodeN 4 PLockUniv
+  icod_ PIntervalUniv = icodeN 5 PIntervalUniv
 
   value = vcase valu where
     valu [0, a] = valuN PType a
@@ -270,6 +273,7 @@ instance EmbPrj NLPSort where
     valu [2, f, a] = valuN PInf f a
     valu [3]    = valuN PSizeUniv
     valu [4]    = valuN PLockUniv
+    valu [5]    = valuN PIntervalUniv
     valu _      = malformed
 
 instance EmbPrj RewriteRule where