+++ src/full/Agda/TypeChecking/SyntacticEquality.hs
@@ -120,6 +120,7 @@ instance SynEq Sort where
       (UnivSort a, UnivSort a') -> UnivSort <$$> synEq a a'
       (SizeUniv, SizeUniv  ) -> pure2 s
       (LockUniv, LockUniv  ) -> pure2 s
+      (IntervalUniv, IntervalUniv) -> pure2 s
       (Prop l  , Prop l'   ) -> Prop <$$> synEq l l'
       (Inf f m , Inf f' n) | f == f', m == n -> pure2 s
       (SSet l  , SSet l'   ) -> SSet <$$> synEq l l'