+++ test/Internal/TypeChecking/Generators.hs
@@ -476,6 +476,7 @@ instance ShrinkC Sort where
     Inf f _    -> []
     SizeUniv   -> []
     LockUniv   -> []
+    IntervalUniv -> []
     PiSort a s1 s2 -> __IMPOSSIBLE__
     FunSort s1 s2 -> __IMPOSSIBLE__
     UnivSort s -> __IMPOSSIBLE__